Add files via upload
This program checks your if you are admitted to college. It checks for your GPA and SAT exam score. If you have a GPA below 1.8 then you are rejected and also if you have an SAT score below 900 then you are also rejected, but other wise you are accepted to college.