Add files via upload
This program is a head and tails generator that I made on my own. Basically the program asks you the amount of games you want to play and you type it in. Then it asks you you to choose multiple times heads as an (h) and tails as a (t). For example let's say you type in "hhtth" and the prediction by the program was heads and you got over 50% heads then you win and the opposite way around. At the end it will print your percentage and if you won or not.