Add files via upload
This is a simple program that takes in a bunch of random grocery items that I made up with random prices and basically calculates and displays the total price on the console. The GroceryList class basically has a bunch of items you can buy at the supermarket. The GroceryItemOrder class requests a purchase of a particular item at a certain quantity. Finally the GroceryClient class adds different items along with their item price and quantity and displays the total cost of the items on the console.